Remove prometheus client from etl dependencies
queue_pusher imports github.com/m-lab/etl/etl and github.com/m-lab/etl/storage
Both packages import metrics, which imports the prometheus client libraries
which now include `syscall` imports, which are forbidden in AppEngine SE. As a
result the queue_pusher build fails.

This change moves etl.CountPanic to `metrics`.
This change moves storage.GetFilename to etl globals, adding a unit test.
