+++ etl/globals.go
@@ -1,16 +1,13 @@
 package etl
 
 import (
+	"encoding/base64"
 	"errors"
-	"fmt"
-	"log"
 	"net"
 	"os"
 	"regexp"
-	"runtime/debug"
 	"strconv"
-
-	"github.com/m-lab/etl/metrics"
+	"strings"
 )
 
 // IsBatch indicates this process is a batch processing service.
@@ -269,60 +266,21 @@ func (dt DataType) Table() string {
 	return dataTypeToTable[dt]
 }
 
-// CountPanics updates the PanicCount metric, then repanics.
-// It must be wrapped in a defer.
-// Examples:
-//  For function that returns an error:
-//    func foobar() () {
-//        defer func() {
-//		      etl.AddPanicMetric(recover(), "foobar")
-// 	      }()
-//        ...
-//        ...
-//    }
-// TODO - possibly move this to metrics.go
-func CountPanics(r interface{}, tag string) {
-	if r != nil {
-		err, ok := r.(error)
-		if !ok {
-			log.Println("bad recovery conversion")
-			err = fmt.Errorf("pkg: %v", r)
-		}
-		log.Println("Adding metrics for panic:", err)
-		metrics.PanicCount.WithLabelValues(tag).Inc()
-		panic(r)
+// GetFilename converts request received from the queue into a filename.
+// TODO(dev) Add unit test
+func GetFilename(filename string) (string, error) {
+	if strings.HasPrefix(filename, "gs://") {
+		return filename, nil
 	}
-}
 
-// PanicToErr captures panics and converts them to
-// errors.  Use with extreme care, as panic may mean that
-// state is corrupted, and continuing to execut may result
-// in undefined behavior.
-// It must be wrapped in a defer.
-// Example:
-//    // err must be a named return value to be captured.
-//    func foobar() (err error) {
-//        defer func() {
-//			  // Possibly do something with existing error
-//            // before calling PanicToErr
-//		      err = etl.PanicToErr(err, recover(), "foobar")
-// 	      }()
-//        ...
-//        ...
-//    }
-func PanicToErr(err error, r interface{}, tag string) error {
-	if r != nil {
-		var ok bool
-		err, ok = r.(error)
-		// TODO - Check if err == runtime.Error, and treat
-		// differently ?
-		if !ok {
-			log.Println("bad recovery conversion")
-			err = fmt.Errorf("pkg: %v", r)
-		}
-		log.Println("Recovered from panic:", err)
-		metrics.PanicCount.WithLabelValues(tag).Inc()
-		fmt.Printf("%s\n", debug.Stack())
+	decode, err := base64.StdEncoding.DecodeString(filename)
+	if err != nil {
+		return "", errors.New("invalid file path: " + filename)
 	}
-	return err
+	fn := string(decode[:])
+	if strings.HasPrefix(fn, "gs://") {
+		return fn, nil
+	}
+
+	return "", errors.New("invalid base64 encoded file path: " + fn)
 }
